rag-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu>=1.8.0
pymupdf==1.24.10
groq>=0.13.0
httpx<0.28
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.8.2
python-multipart>=0.0.9
